+++ examples/wasm/test/BUILD
@@ -1,0 +1,10 @@
+cc_test(
+    name = "config_test",
+    srcs = ["config_test.cc"],
+    deps = [
+        "//examples/wasm/utility:config",
+        "@gtest//:gtest",
+        "@gtest//:gtest_main"
+    ],
+)
+
