+++ examples/wasm/utility/common.h
@@ -1,0 +1,13 @@
+// NOLINT(namespace-envoy)
+#include <string>
+#include <vector>
+#include <unordered_map>
+#include <unordered_set>
+
+#pragma once
+
+using Keys = std::unordered_set<std::string>;
+using QueryParams = std::unordered_map<std::string, std::string>;
+const std::string URLENCODED = "application/x-www-form-urlencoded";
+
+
